pyproject.toml
[project]
name = "PalworldSavePal"
version = "0.14.1-b1"
requires-python = ">=3.12"
dependencies = [
    "altgraph==0.17.4",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "bottle==0.13.2",
    "cffi==1.17.1",
    "click==8.1.7",
    "clr-loader==0.2.6",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "cx-freeze==7.2.7",
    "deepdiff==8.0.1",
    "fastapi==0.115.6",
    "h11==0.16.0",
    "idna==3.10",
    "lief==0.15.1",
    "ordered-set==4.1.0",
    "orderly-set==5.2.2",
    "packaging==25.0",
    "pefile==2023.2.7",
    "proxy-tools==0.1.0",
    "psutil==6.1.0",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pythonnet==3.0.4",
    "pywebview==5.4",
    "pywin32-ctypes==0.2.3",
    "setuptools==75.6.0",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "sqlalchemy==2.0.38",
    "sqlmodel==0.0.23",
    "typing-extensions==4.12.2",
    "uvicorn==0.32.1",
    "websockets==15.0.1",
    "zstandard==0.23.0",
    "palworld-save-tools @ git+https://github.com/oMaN-Rod/palworld-save-tools.git@main",
    "PyGObject==3.52.3; sys_platform == 'linux'",
]

[tool.uv]
package = false
